% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_simulation_master}
\alias{run_simulation_master}
\title{Shiny App of Simulation Master}
\usage{
run_simulation_master()
}
\description{
This is a function to run the Shiny App of the "Simulation Master".
This Shiny App contains two menus, one of which is regarding regression:
it provides different parameters to generate sparse data, and also parameters to tune the Lasso/Ridge/Elastic Net;
the other of which is regarding classification: it provides the generation functionality of
four types of data, and tuning parameters of kernel SVM as well. This Shiny App focus on the
feature selection of regression problems and the prediction performance of classification problems.
}
\note{
This is simply a function to trigger the Shiny App, so it bears no parameters or values.
}
\examples{
\dontrun{
run_simulation_master()
}

}
